// src/app/core/models/statistique/eleve/dashboard-stats-eleve.ts
export interface DashboardStatsEleve {
  moyenneGeneraleEleve?: number;
  classementGeneralEleve?: number;
  nombreAbsenceEleve?: number;
  nombreAbsenceJustifieEleve?: number;
  nombreRetardEleve?: number;
  montantTotalEnAttanteEleve?: number;
  montantTotalEnRetardEleve?: number;
  montantTotalImpayeEleve?: number;
}